+++ Application/DiSpeakBouyomiChanBridge/CommandSystem/Command/Abstract/ExecutableCommand.cs
@@ -1,13 +1,12 @@
-﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem
+﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem;
+
+/// <summary>
+/// 文字列中から検出する実行可能なコマンドの基底クラス
+/// </summary>
+public abstract class ExecutableCommand
 {
     /// <summary>
-    /// 文字列中から検出する実行可能なコマンドの基底クラス
+    /// 処理を実行します
     /// </summary>
-    public abstract class ExecutableCommand
-    {
-        /// <summary>
-        /// 処理を実行します
-        /// </summary>
-        public abstract void Execute();
-    }
+    public abstract void Execute();
 }
