+++ Application/DiSpeakBouyomiChanBridge/CommandSystem/Command/Abstract/ExecutableCommandFactory.cs
@@ -2,28 +2,27 @@
 
 using net.boilingwater.Framework.Core;
 
-namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem
+namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem;
+
+/// <summary>
+/// <see cref="ExecutableCommand"/>を文字列中から作成するファクトリの基底クラス
+/// </summary>
+public abstract class ExecutableCommandFactory
 {
     /// <summary>
-    /// <see cref="ExecutableCommand"/>を文字列中から作成するファクトリの基底クラス
+    /// コマンドを保持する辞書
     /// </summary>
-    public abstract class ExecutableCommandFactory
-    {
-        /// <summary>
-        /// コマンドを保持する辞書
-        /// </summary>
-        public SimpleDic<ExecutableCommand> Dic { get; protected set; } = [];
+    public SimpleDic<ExecutableCommand> Dic { get; protected set; } = [];
 
-        /// <summary>
-        /// コマンドファクトリの初期化を行います
-        /// </summary>
-        public abstract void Initialize();
+    /// <summary>
+    /// コマンドファクトリの初期化を行います
+    /// </summary>
+    public abstract void Initialize();
 
-        /// <summary>
-        /// コマンドを検出します。
-        /// </summary>
-        /// <param name="input">コマンドを含む文字列<br/>検出したコマンドは文字列中から削除されます</param>
-        /// <returns>文字列中から検出した<seealso cref="IEnumerable{ExecutableCommand}"/></returns>
-        public abstract IEnumerable<ExecutableCommand> CreateExecutableCommands(ref string input);
-    }
+    /// <summary>
+    /// コマンドを検出します。
+    /// </summary>
+    /// <param name="input">コマンドを含む文字列<br/>検出したコマンドは文字列中から削除されます</param>
+    /// <returns>文字列中から検出した<seealso cref="IEnumerable{ExecutableCommand}"/></returns>
+    public abstract IEnumerable<ExecutableCommand> CreateExecutableCommands(ref string input);
 }
