+++ Application/DiSpeakBouyomiChanBridge/CommandSystem/Handler/Impl/DiSpeakCommandHandler.cs
@@ -1,6 +1,5 @@
-﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem.Handle.Impl
+﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem.Handle.Impl;
+
+internal class DiSpeakCommandHandler : AbstractCommandHandler
 {
-    internal class DiSpeakCommandHandler : AbstractCommandHandler
-    {
-    }
 }
