+++ Application/DiSpeakBouyomiChanBridge/CommandSystem/Handler/Impl/InternalDiscordClientCommandHandler.cs
@@ -1,6 +1,5 @@
-﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem.Handle.Impl
+﻿namespace net.boilingwater.Application.DiSpeakBouyomiChanBridge.CommandSystem.Handle.Impl;
+
+internal class InternalDiscordClientCommandHandler : AbstractCommandHandler
 {
-    internal class InternalDiscordClientCommandHandler : AbstractCommandHandler
-    {
-    }
 }
